pyproject.toml
[project]
name = "ai-usecase-agents"
version = "0.1.0"
description = "Multi-agent system to generate GenAI use cases for a company or industry"
authors = [
    {name = "Rishabh"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai (>=0.114.0,<0.115.0)",
    "tavily-python (>=0.5.4,<0.6.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
